Refactor fully
All line detection and splitting is now done in a separate LineBuilder
class. Lines which require OCR are later picked up by OCRBuilder
